+++ src/app/loadout/ingame/selectors.ts
@@ -18,8 +18,8 @@ import {
 } from 'app/loadout-drawer/loadout-types';
 import {
   getModsFromLoadout,
+  itemsByItemId,
   newLoadoutFromEquipped,
-  potentialLoadoutItemsByItemId,
 } from 'app/loadout-drawer/loadout-utils';
 import { loadoutsForClassTypeSelector } from 'app/loadout-drawer/loadouts-selector';
 import { d2ManifestSelector } from 'app/manifest/selectors';
@@ -146,7 +146,7 @@ export const inGameLoadoutsWithMetadataSelector = createSelector(
         .filter((gameLoadout) => gameLoadout.index < availableLoadoutSlots)
         .map((gameLoadout) => {
           const isEquippable = gameLoadout.items.every((li) => {
-            const liveItem = potentialLoadoutItemsByItemId(allItems)[li.itemInstanceId];
+            const liveItem = itemsByItemId(allItems)[li.itemInstanceId];
             return !liveItem || itemCouldBeEquipped(selectedStore, liveItem, stores);
           });
 
@@ -157,15 +157,13 @@ export const inGameLoadoutsWithMetadataSelector = createSelector(
             currentLoadout.resolvedMods,
           );
 
-          const matchingLoadouts = savedLoadouts.filter(
-            (dimLoadout) =>
-              dimLoadout.loadout.items.length > 4 &&
-              implementsDimLoadout(
-                defs,
-                gameLoadout,
-                dimLoadout.resolvedLoadoutItems,
-                dimLoadout.resolvedMods,
-              ),
+          const matchingLoadouts = savedLoadouts.filter((dimLoadout) =>
+            implementsDimLoadout(
+              defs,
+              gameLoadout,
+              dimLoadout.resolvedLoadoutItems,
+              dimLoadout.resolvedMods,
+            ),
           );
           return { gameLoadout, isEquippable, isEquipped, matchingLoadouts };
         })